Move tests for main_install.py module into a separate cli/test_main_install.py file
